Removed files no longer in use
