+++ CMakeLists.txt
@@ -60,7 +60,6 @@ add_compile_options(-Wall -Wconversion -Wextra -pedantic)
 include_directories(${TOPICS_DIR})
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include/fastdds)
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include/application)
-# include_directories(/home/pi/fastdds_python_ws/install/fastdds/include)
 create_msg_library(ToFData)
 create_msg_library(MotorCommands)
 
@@ -75,48 +74,15 @@ set(TOF_SOURCES
     ${DRIVER_DIR}/sensors/tof_sensor.cpp
     )
 
-# # Add executable for publisher
-# set(SOURCES
-#     src/main_combined.cpp
-#     ${TOF_SOURCES}
-#     ${APPLICATION_DIR}/ToFDataPublisher.cpp
-#     ${APPLICATION_DIR}/MotorMove.cpp
-#     )
-
-set(TEST_SOURCES
-    src/tester_main.cpp
+set(MAIN_SOURCES
+    src/main.cpp
     ${APPLICATION_DIR}/MotorMove.cpp
     ${APPLICATION_DIR}/DDSCommunicator.cpp
     ${TOF_SOURCES}
 )
 
-add_executable(TofTest ${TEST_SOURCES})
+add_executable(SmartWalkingStick ${MAIN_SOURCES})
 
-target_link_libraries(TofTest ${LIBGPIO_LIBRARIES} gpiod  fastcdr fastdds ToFData MotorCommands)
+target_link_libraries(SmartWalkingStick ${LIBGPIO_LIBRARIES} gpiod fastcdr fastdds ToFData MotorCommands)
 
 
-# message(STATUS "SUB_SOURCES: ${SUB_SOURCES}")
-# add_executable(DDSPublisher ${PUB_SOURCES})
-# target_link_libraries(DDSPublisher fastcdr fastrtps ToFData MotorCommands)
-
-# add_executable(DDSSubscriber ${SUB_SOURCES})
-# target_link_libraries(DDSSubscriber fastcdr fastrtps ToFData MotorCommands)
-#
-
-# add_executable(CombinedSystem ${SOURCES})
-# target_link_libraries(CombinedSystem fastcdr fastrtps ToFData MotorCommands ${LIBGPIO_LIBRARIES})
-
-# set(SOURCES
-#     main.cpp
-#     ${DRIVER_DIR}/sensors/sensors.cpp
-#     ${DRIVER_DIR}/sensors/tof_sensor.cpp)    
-
-# set(TOF_SOURCES
-#     ${DRIVER_DIR}/vl53l0x-linux/VL53L0X.cpp
-#     ${DRIVER_DIR}/vl53l0x-linux/I2Cdev.cpp
-#     )
-
-# add_executable(SensorApp ${SOURCES} ${TOF_SOURCES})
-# target_link_libraries(SensorApp ${LIBGPIO_LIBRARIES})
-
-# add_subdirectory(src/ToF_GUI)
